feat: DTO 객체 하나로 통일 (#15)
- record로 관리
