--- service/src/main/java/way/application/service/schedule/dto/request/DeleteScheduleRequestDto.java
@@ -1,7 +1,0 @@
-package way.application.service.schedule.dto.request;
-
-public record DeleteScheduleRequestDto(
-	Long scheduleSeq,
-	Long creatorSeq
-) {
-}
