--- service/src/main/java/way/application/service/schedule/dto/response/ModifyScheduleResponseDto.java
@@ -1,7 +1,0 @@
-package way.application.service.schedule.dto.response;
-
-public record ModifyScheduleResponseDto(
-	Long scheduleSeq
-) {
-
-}
