+++ service/src/main/java/way/application/service/schedule/mapper/ScheduleMapper.java
@@ -6,10 +6,10 @@ import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
 
 import way.application.infrastructure.schedule.entity.ScheduleEntity;
-import way.application.service.schedule.dto.request.SaveScheduleRequestDto;
+import way.application.service.schedule.dto.request.ScheduleRequestDto;
 
 @Mapper(componentModel = "spring", injectionStrategy = InjectionStrategy.CONSTRUCTOR, unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface ScheduleMapper {
 	@Mapping(target = "scheduleSeq", ignore = true)
-	ScheduleEntity toScheduleEntity(SaveScheduleRequestDto scheduleDto);
+	ScheduleEntity toScheduleEntity(ScheduleRequestDto.SaveScheduleRequestDto scheduleDto);
 }
