+++ service/src/main/java/way/application/service/schedule/service/ScheduleService.java
@@ -17,12 +17,8 @@ import way.application.infrastructure.schedule.entity.ScheduleEntity;
 import way.application.infrastructure.schedule.repository.ScheduleRepository;
 import way.application.infrastructure.scheduleMember.entity.ScheduleMemberEntity;
 import way.application.infrastructure.scheduleMember.repository.ScheduleMemberRepository;
-import way.application.service.schedule.dto.request.DeleteScheduleRequestDto;
-import way.application.service.schedule.dto.request.ModifyScheduleRequestDto;
-import way.application.service.schedule.dto.request.SaveScheduleRequestDto;
-import way.application.service.schedule.dto.response.GetScheduleResponseDto;
-import way.application.service.schedule.dto.response.ModifyScheduleResponseDto;
-import way.application.service.schedule.dto.response.SaveScheduleResponseDto;
+import way.application.service.schedule.dto.request.ScheduleRequestDto;
+import way.application.service.schedule.dto.response.ScheduleResponseDto;
 import way.application.service.schedule.mapper.ScheduleMapper;
 import way.application.service.scheduleMember.mapper.ScheduleMemberMapper;
 
@@ -48,9 +44,10 @@ public class ScheduleService {
 	 * Service Layer -> Repo의 CRUD만 처리
 	 */
 
-
 	@Transactional
-	public SaveScheduleResponseDto createSchedule(SaveScheduleRequestDto saveScheduleRequestDto) {
+	public ScheduleResponseDto.SaveScheduleResponseDto createSchedule(
+		ScheduleRequestDto.SaveScheduleRequestDto saveScheduleRequestDto
+	) {
 		// Member 유효성 검사 (Repository 에서 처리)
 		MemberEntity createMemberEntity = memberRepository.validateMemberSeq(saveScheduleRequestDto.createMemberSeq());
 		List<MemberEntity> invitedMemberEntity = memberRepository.validateMemberSeqs(
@@ -77,11 +74,13 @@ public class ScheduleService {
 				firebaseNotificationDomain.sendNotification(invitedMember, createMemberEntity);
 		}
 
-		return new SaveScheduleResponseDto(savedSchedule.getScheduleSeq());
+		return new ScheduleResponseDto.SaveScheduleResponseDto(savedSchedule.getScheduleSeq());
 	}
 
 	@Transactional
-	public ModifyScheduleResponseDto modifySchedule(ModifyScheduleRequestDto modifyScheduleRequestDto) {
+	public ScheduleResponseDto.ModifyScheduleResponseDto modifySchedule(
+		ScheduleRequestDto.ModifyScheduleRequestDto modifyScheduleRequestDto
+	) {
 		// 유효성 검사 (Repository 에서 처리)
 		MemberEntity createMemberEntity = memberRepository.validateMemberSeq(
 			modifyScheduleRequestDto.createMemberSeq()
@@ -89,6 +88,7 @@ public class ScheduleService {
 		List<MemberEntity> invitedMemberEntity = memberRepository.validateMemberSeqs(
 			modifyScheduleRequestDto.invitedMemberSeqs()
 		);
+		ScheduleEntity savedSchedule = scheduleRepository.validateScheduleSeq(modifyScheduleRequestDto.scheduleSeq());
 		ScheduleEntity scheduleEntity = scheduleMemberRepository.validateScheduleEntityCreatedByMember(
 			modifyScheduleRequestDto.scheduleSeq(),
 			modifyScheduleRequestDto.createMemberSeq()
@@ -99,15 +99,15 @@ public class ScheduleService {
 		scheduleMemberRepository.deleteAllBySchedule(scheduleEntity);
 
 		// 재저장
-		SaveScheduleResponseDto saveScheduleResponseDto = createSchedule(
+		ScheduleResponseDto.SaveScheduleResponseDto saveScheduleResponseDto = createSchedule(
 			modifyScheduleRequestDto.toSaveScheduleRequestDto()
 		);
 
-		return new ModifyScheduleResponseDto(saveScheduleResponseDto.scheduleSeq());
+		return new ScheduleResponseDto.ModifyScheduleResponseDto(saveScheduleResponseDto.scheduleSeq());
 	}
 
 	@Transactional
-	public void deleteSchedule(DeleteScheduleRequestDto deleteScheduleRequestDto) {
+	public void deleteSchedule(ScheduleRequestDto.DeleteScheduleRequestDto deleteScheduleRequestDto) {
 		// 유효성 검사 (Repository 에서 처리)
 		memberRepository.validateMemberSeq(deleteScheduleRequestDto.creatorSeq());
 		scheduleRepository.validateScheduleSeq(deleteScheduleRequestDto.scheduleSeq());
@@ -123,7 +123,7 @@ public class ScheduleService {
 
 	@Cacheable(value = "schedules", key = "#scheduleSeq + '-' + #memberSeq")
 	@Transactional(readOnly = true)
-	public GetScheduleResponseDto getSchedule(Long scheduleSeq, Long memberSeq) {
+	public ScheduleResponseDto.GetScheduleResponseDto getSchedule(Long scheduleSeq, Long memberSeq) {
 		// 유효성 검사 (Repository 에서 처리)
 		memberRepository.validateMemberSeq(memberSeq);
 		ScheduleEntity scheduleEntity = scheduleRepository.validateScheduleSeq(scheduleSeq);
@@ -137,7 +137,7 @@ public class ScheduleService {
 		// userName 추출 (Domain Layer)
 		List<String> userName = scheduleMemberDomain.extractUserNameFromScheduleMemberEntities(scheduleEntities);
 
-		return new GetScheduleResponseDto(
+		return new ScheduleResponseDto.GetScheduleResponseDto(
 			scheduleEntity.getTitle(), scheduleEntity.getStartTime(), scheduleEntity.getEndTime(),
 			scheduleEntity.getLocation(), scheduleEntity.getStreetName(), scheduleEntity.getX(), scheduleEntity.getY(),
 			scheduleEntity.getColor(), scheduleEntity.getMemo(), userName
